Create Group Anagrams (Using Hashmap) - GoldmanSachs
This is the second approach to the question : Given an array of strings strs, group the anagrams together. You can return the answer in any order. Link : https://leetcode.com/problems/group-anagrams/

The idea is simple : The key of hashmap is a string and value is an arraylist containing group anagrams.
Suppose the input string is ["eat","ate","drink","tea","kind"], then the key will contain string with sorted characters and value would be an arraylist of anagrams corresponding to that sorted string. 
Key : "aet"
Value corresponding to this key  ArrayList : ["eat","ate","tea"]
The time complexity of this approach is O(n).